requirements.txt
# Core dependencies
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM & AI
groq==0.4.1
langchain==0.1.0
langchain-community==0.0.13
langchain-groq==0.0.1

# Vector Store
supabase==2.3.0
pgvector==0.2.4
psycopg2-binary==2.9.9

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
python-multipart==0.0.6

# Text Processing
tiktoken==0.5.2
sentence-transformers==2.2.2

# Frontend
streamlit==1.29.0
plotly==5.18.0
streamlit-option-menu==0.3.6

# Utilities
requests==2.31.0
aiohttp==3.9.1
python-json-logger==2.0.7
